Refactorización de pestañas
Refactorización del nombre de algunas pestañas y controladores para adaptarlos a la nueva estructura
