+++ AlimentApp/src/main/java/dad/alimentapp/service/ProfileService.java
@@ -4,6 +4,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+import dad.alimentapp.controllers.ProfileController;
 import dad.alimentapp.main.App;
 import dad.alimentapp.models.Gender;
 import dad.alimentapp.models.Profile;
@@ -42,4 +43,28 @@ public class ProfileService {
 		}
 		return profile;
 	}
+	
+	
+	
+	/**
+	 * Método que carga la lista de perfiles guardados en la base de datos
+	 */
+		public static void loadProfiles() {
+			
+			ProfileController.profileList.clear();
+
+			try {
+				String sql = "SELECT * FROM profile";
+				PreparedStatement query = App.connection.prepareStatement(sql);
+				ResultSet result = query.executeQuery();
+				while (result.next()) {
+					ProfileController.profileList.add(new Profile(result.getInt(1), result.getString(2), result.getString(3),
+							result.getString(4), result.getInt(5), result.getInt(6), result.getInt(7), result.getDouble(8),
+							Gender.valueOf(result.getInt(9))));
+				}
+			} catch (SQLException e) {
+				Messages.error("Error al obtener la lista de perfiles", e.getMessage());
+			}
+
+		}
 }
